+++ Editor/Core.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: ac4a2567e8440574090ea5eb33cc3da4
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
