+++ Editor/Core/ILogFileSlicer.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: d31f003965322424d957e552f5146d06
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
