+++ Editor/Core/LogDataFileSlicer.cs
