+++ Editor/Core/RawDataFileSlicer.cs
