+++ Editor/UI/UXML/ProfilerLogSplit.uxml.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: cd8532e7b2b56b7489a3d87c0a54911a
+ScriptedImporter:
+  fileIDToRecycleName:
+    11400000: tree
+    11400002: inlineStyle
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant: 
+  script: {fileID: 13804, guid: 0000000000000000e000000000000000, type: 0}
