+++ Editor/UI/UXML2018/ProfilerLogSplit.uxml
@@ -1,0 +1,18 @@
+<ui:UXML xmlns:ui="UnityEngine.Experimental.UIElements" xmlns:uie="UnityEditor.Experimental.UIElements">
+    <ui:Label text="Select Profiler Log File" />
+    <ui:VisualElement style="flex-wrap: nowrap;
+ flex-direction: row;
+ margin-left: 20px;">
+        <ui:TextField picking-mode="Ignore" value="filler text" name="TextField" style="flex-grow: 1;" />
+        <ui:Button text="File" name="Button" style="width: 60px;" />
+    </ui:VisualElement>
+    <ui:VisualElement name="AfterAnalyze" style="margin-top: 10px;
+ margin-left: 20px;
+ margin-right: 10px;">
+        <ui:Label text="Label" />
+        <ui:MinMaxSlider picking-mode="Ignore" label="ProfilerFrame" min-value="10" max-value="12" low-limit="-10" high-limit="40" name="ProfilerFrame" style="padding-left: 5px;" />
+        <ui:TextField picking-mode="Ignore" label="MinFrame" value="filler text" name="ProfilerMinFrame" style="padding-left: 5px;" />
+        <ui:TextField picking-mode="Ignore" label="MaxFrame" value="filler text" name="ProfilerMaxFrame" style="padding-left: 5px;" />
+        <ui:Button text="SendToProfiler" name="SendProfilerBtn" style="width: 160px;" />
+    </ui:VisualElement>
+</ui:UXML>
